basic bandwidth-based server rate limiting
So far this server is not performing as well as the node-based one, but it is passable. I couldn't use the same rate-limiting method as for the client, because we don't get bytesWritten() signals on the server connections. I've tuned the example config to work well locally, on my machine.